oneshots/openstack-lister/requirements.txt
flask
openstacksdk
cached_property
